[refactor] Game class -> GameClient class 로 변경
