+++ src/main/java/racingcar/Application.java
@@ -2,6 +2,6 @@ package racingcar;
 
 public class Application {
     public static void main(String[] args) {
-        new Game().play();
+        new GameClient().play();
     }
 }
