+++ src/main/java/racingcar/GameClient.java
@@ -1,0 +1,19 @@
+package racingcar;
+
+import racingcar.controller.GameController;
+
+public class GameClient {
+
+	private GameController gameController;
+
+	public GameClient() {
+		this.gameController = new GameController();
+	}
+
+	public void play() {
+		gameController.inputCars();
+		gameController.inputRepeatNumber();
+		gameController.postResult();
+		gameController.postWinner();
+	}
+}
