+++ src/main/java/racingcar/message/dto/MessageResponse.java
@@ -1,12 +1,8 @@
-package racingcar.message;
+package racingcar.message.dto;
 
 public class MessageResponse {
 
 	public static void of(String message){
 		System.out.println(message);
-	}
-
-	public static void NonLineOf(String message){
-		System.out.print(message);
 	}
 }
